==> Analysis/scripts/00_main.R <==
## Initialize Project




# Load Packages -----------------------------------------------------------

source("./packages/packages.R")
      # Functions Require packages (e.g. dplyr)


# Source Functions --------------------------------------------------------

source("./source/loader.R")
